Add files via upload
This file contains few examples of SQL execution techniques using Python commands. Libraries such as pandas, seaborn, matplotlib, csv and sqlite3 were incorporated to this small project to upload, manipulate and plot the results. This examples are vary useful for those people who want to see the application of SQL and Python in real-world Examples.