Add files via upload
This file contains a good example of creating tables, loading and querying of data using Python pandas and mysql.connector  libraries. So far, we connect to local database using mysql.connector library and create a table using cursor. Later, we insert data to that table and retrieve it using pandas or fetchall  commands.